+++ packages/mobile/src/screens/edit-track-screen/EditTrackNavigator.tsx
@@ -7,10 +7,10 @@ import { useAppScreenOptions } from 'app/screens/app-screen/useAppScreenOptions'
 import { messages as completeMessages } from '../upload-screen/screens/CompleteTrackScreen'
 
 import { EditTrackForm } from './EditTrackForm'
-import { accessAndSaleScreenName } from './fields'
+import { priceAndAudienceScreenName } from './fields'
 import { VisibilityScreen } from './fields/VisibilityField'
 import {
-  AccessAndSaleScreen,
+  PriceAndAudienceScreen,
   AdvancedOptionsScreen,
   IsrcIswcScreen,
   LicenseTypeScreen,
@@ -60,8 +60,8 @@ export const EditTrackNavigator = (props: EditTrackNavigatorProps) => {
           component={AdvancedOptionsScreen}
         />
         <Stack.Screen
-          name={accessAndSaleScreenName}
-          component={AccessAndSaleScreen}
+          name={priceAndAudienceScreenName}
+          component={PriceAndAudienceScreen}
         />
         <Stack.Screen name='NFTCollections' component={NFTCollectionsScreen} />
         <Stack.Screen name='IsrcIswc' component={IsrcIswcScreen} />
