--- packages/mobile/src/screens/edit-track-screen/fields/AccessAndSaleField/index.ts
@@ -1,1 +1,0 @@
-export * from './AccessAndSaleField'
