+++ packages/mobile/src/screens/edit-track-screen/fields/PriceAndAudienceField/PriceAndAudienceField.tsx
@@ -13,7 +13,7 @@ import { useField } from 'formik'
 import type { ContextualMenuProps } from 'app/components/core'
 import { ContextualMenu } from 'app/components/core'
 
-export const accessAndSaleScreenName = 'AccessAndSale'
+export const priceAndAudienceScreenName = 'PriceAndAudience'
 
 const messages = {
   accessAndSale: 'Access & Sale',
@@ -41,9 +41,9 @@ const fieldVisibilityLabelMap = {
 
 const fieldVisibilityKeys = Object.keys(fieldVisibilityLabelMap)
 
-type AccessAndSaleFieldProps = Partial<ContextualMenuProps>
+type PriceAndAudienceFieldProps = Partial<ContextualMenuProps>
 
-export const AccessAndSaleField = (props: AccessAndSaleFieldProps) => {
+export const PriceAndAudienceField = (props: PriceAndAudienceFieldProps) => {
   const [{ value: streamConditions }] =
     useField<Nullable<AccessConditions>>('stream_conditions')
   const [{ value: isUnlisted }] = useField<boolean>('is_unlisted')
@@ -81,7 +81,7 @@ export const AccessAndSaleField = (props: AccessAndSaleFieldProps) => {
   return (
     <ContextualMenu
       label={messages.accessAndSale}
-      menuScreenName={accessAndSaleScreenName}
+      menuScreenName={priceAndAudienceScreenName}
       value={trackAvailabilityLabels}
       {...props}
     />
