+++ packages/mobile/src/screens/edit-track-screen/fields/PriceAndAudienceField/index.ts
@@ -1,0 +1,1 @@
+export * from './PriceAndAudienceField'
