+++ packages/mobile/src/screens/edit-track-screen/fields/index.ts
@@ -7,7 +7,7 @@ export * from './RemixSettingsField'
 export * from './ReleaseDateField'
 export * from './ReleaseDateFieldLegacy'
 export * from './AdvancedOptionsField'
-export * from './AccessAndSaleField'
+export * from './PriceAndAudienceField'
 export * from './SwitchField'
 export * from './IsrcIswcField'
 export * from './LicenseTypeField'
