+++ packages/mobile/src/screens/edit-track-screen/screens/index.ts
@@ -2,7 +2,7 @@ export * from './SelectGenreScreen'
 export * from './SelectMoodScreen'
 export * from './RemixSettingsScreen'
 export * from './AdvancedOptionsScreen'
-export * from './AccessAndSaleScreen'
+export * from './PriceAndAudienceScreen'
 export * from './IsrcIswcScreen'
 export * from './LicenseTypeScreen'
 export * from './ReleaseDateScreen'
